refactor: decompoe Ajustes.vue em View e subcomponentes
- Move a lógica de estado, watchers e chamadas de API para views/AjustesView.vue
- Extrai configurações de horário e dias para components/ajustes/AjustesFuncionamento.vue
- Isola a gestão de capacidades e grupos em components/ajustes/AjustesCapacidade.vue
- Cria components/ajustes/AjustesTags.vue para o gerenciamento de tags
- Move o card de navegação externa para components/ajustes/AjustesProtocolosLink.vue
- Atualiza o roteador para utilizar a nova AjustesView

Fixes #46
